+++ src/Resources/contao/dca/tl_ctv_inserttag.php
@@ -88,7 +88,7 @@ $GLOBALS['TL_DCA']['tl_ctv_inserttag'] = [
             .'{ctv_legend},ctvDatei,ctvFormat,ctvListe;'
             .'{ctv_spalten_legend},ctvSpalten;'
             .'{ctv_runden_legend},ctvStand,ctvRunden;'
-            .'{ctv_ueberschrift_legend},ctvStandAus,ctvRundenkopfAus;'
+            .'{ctv_ueberschrift_legend},ctvUeberschriftenAus;'
             .'{ctv_mannschaft_legend},ctvMannschaftswahl,ctvMannschaftSpieler,ctvKreuzKurz;'
             .'{ctv_hinweis_legend},ctvDatum,ctvHinweise',
     ],
@@ -139,7 +139,7 @@ $GLOBALS['TL_DCA']['tl_ctv_inserttag'] = [
  * Inhaltselements aus der Zeit vor Fassung 1.8.0 und hat in einer neuen
  * Tabelle nichts zu suchen.
  */
-foreach (['ctvDatei', 'ctvFormat', 'ctvListe', 'ctvSpalten', 'ctvStand', 'ctvStandAus', 'ctvRunden', 'ctvRundenkopfAus', 'ctvMannschaftswahl', 'ctvMannschaftSpieler', 'ctvKreuzKurz', 'ctvDatum', 'ctvHinweise'] as $feld) {
+foreach (['ctvDatei', 'ctvFormat', 'ctvListe', 'ctvSpalten', 'ctvStand', 'ctvRunden', 'ctvUeberschriftenAus', 'ctvStandAus', 'ctvRundenkopfAus', 'ctvMannschaftswahl', 'ctvMannschaftSpieler', 'ctvKreuzKurz', 'ctvDatum', 'ctvHinweise'] as $feld) {
     if (isset($GLOBALS['TL_DCA']['tl_content']['fields'][$feld])) {
         $GLOBALS['TL_DCA']['tl_ctv_inserttag']['fields'][$feld] = $GLOBALS['TL_DCA']['tl_content']['fields'][$feld];
     }